Added Camera class, cleaned up and streamlined code for drawing shapes
Instead of the TriangleShape handling most of the drawing code, the code has been split up between the Window and Shape class. The user calls window.draw(shape), then the shape calls the corresponding draw function from the window, i.e. window.drawShape(). With this function, the shape passes all of the necessary data for drawing. With this implementation, the majority of drawing code is bundled into the Window class, making it easy to understand what's going on. Also, we won't have to create a shader program for every single shape object, as the window uses its own. Later on, we can also add additional draw functions to the window class, e.g. window.drawSprite, window.drawText.